pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "content-chatbot"
version = "0.1.0"
description = "LangGraph-based chatbot with Streamlit frontend"
requires-python = ">=3.9"
dependencies = [
    "streamlit==1.40.2",
    "langchain",
    "langchain-openai",
    "langgraph",
    "langgraph-checkpoint-sqlite",
    "python-dotenv",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "frontend*"]
